% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{calculate_lambda}
\alias{calculate_lambda}
\title{Calculate Lambda}
\usage{
calculate_lambda(temperature, humidity, pressure)
}
\arguments{
\item{temperature}{A numeric value representing the temperature.}

\item{humidity}{A numeric value representing the humidity percentage.}

\item{pressure}{A numeric value representing the pressure.}
}
\value{
A numeric value representing the calculated lambda.
}
\description{
This function calculates the lambda value based on temperature, humidity, and pressure.
}
\details{
Calculate lambda based on weather
}
\examples{
\dontrun{
  lambda <- calculate_lambda(25, 60, 1013)
  print(lambda)
}
}
