% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_functions.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate Metrics}
\usage{
calculate_metrics(random_variations, x_values)
}
\arguments{
\item{random_variations}{A numeric vector of random variations.}

\item{x_values}{A numeric vector of x-values.}
}
\value{
A list containing the volume and surface area of the cone.
}
\description{
This function calculates the volume and surface area of a cone based on random variations and x-values.
}
\details{
Calculate volume and surface area
}
\examples{
\dontrun{
  metrics <- calculate_metrics(generate_random_variations(1:10), 1:10)
  print(metrics)
}
}
