% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_city_validity}
\alias{check_city_validity}
\title{Check City Validity}
\usage{
check_city_validity(city_name, API_KEY)
}
\arguments{
\item{city_name}{A character string representing the name of the city.}

\item{API_KEY}{A character string representing the OpenWeatherMap API key.}
}
\value{
A logical value (`TRUE` if the city is valid, `FALSE` otherwise).
}
\description{
This function checks if the city name provided is valid by making a request
to the OpenWeatherMap API. It returns `TRUE` if the city is valid, and `FALSE` otherwise.
}
