% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_functions.R
\name{generate_random_variations}
\alias{generate_random_variations}
\title{Generate Random Variations}
\usage{
generate_random_variations(x_values, sd = 0.1)
}
\arguments{
\item{x_values}{A numeric vector of x-values.}

\item{sd}{A numeric value representing the standard deviation for the random variations. Default is 0.1.}
}
\value{
A numeric vector of random variations.
}
\description{
This function generates random variations for cone radius based on the provided x-values.
}
\details{
Generate random variations for cone radius
}
\examples{
\dontrun{
  variations <- generate_random_variations(1:10)
  print(variations)
}
}
