% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_current_weather}
\alias{get_current_weather}
\title{Get Current Weather}
\usage{
get_current_weather(city)
}
\arguments{
\item{city}{A character string representing the name of the city.}
}
\value{
A tibble containing the current weather information for the specified city.
}
\description{
This function retrieves the current weather for a specified city from the OpenWeatherMap API.
The data is returned in a tidy format with various weather-related variables.
}
