% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_module.R
\name{outputModuleUI}
\alias{outputModuleUI}
\title{Output Module UI}
\usage{
outputModuleUI(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\value{
A tagList containing the UI components.
}
\description{
This function creates the UI components for the output module.
}
