% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot.forecast}
\alias{plot.forecast}
\title{Plot Forecast Data}
\usage{
\method{plot}{forecast}(x, variable, ...)
}
\arguments{
\item{x}{A `forecast` object containing the weather forecast data.}

\item{variable}{A character string indicating the variable to be plotted ("temperature", "humidity", or "pressure").}

\item{...}{Additional arguments to be passed to the `ggplot` function.}
}
\value{
A `ggplot` object displaying the weather forecast plot.
}
\description{
This function generates a plot for the 5-day weather forecast of a chosen variable (temperature, humidity, or pressure).
The plot includes a line, points, and an area shaded under the curve.
}
