% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_cones_per_guest}
\alias{simulate_cones_per_guest}
\title{Simulate Cones Per Guest}
\usage{
simulate_cones_per_guest(guest_count)
}
\arguments{
\item{guest_count}{An integer representing the number of guests.}
}
\value{
An integer representing the total number of cones consumed.
}
\description{
This function simulates the number of cones consumed per guest.
}
\details{
Simulate cone consumption per guest
}
\examples{
\dontrun{
  cones <- simulate_cones_per_guest(10)
  print(cones)
}
}
