% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_guest_count}
\alias{simulate_guest_count}
\title{Simulate Guest Count}
\usage{
simulate_guest_count(lambda)
}
\arguments{
\item{lambda}{A numeric value representing the lambda parameter for the Poisson distribution.}
}
\value{
An integer representing the simulated number of guests.
}
\description{
This function simulates the number of guests based on the lambda value.
}
\details{
Simulate the number of guests based on lambda
}
\examples{
\dontrun{
  guest_count <- simulate_guest_count(5)
  print(guest_count)
}
}
