% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_weather_forecast}
\alias{get_weather_forecast}
\title{Get Weather Forecast}
\usage{
get_weather_forecast(city, API_KEY)
}
\arguments{
\item{city}{A character string representing the name of the city.}
}
\value{
A dataframe containing the 5-day weather forecast for the specified city.
}
\description{
This function retrieves a 5-day forecast (every 3 hours) for a specified city from the OpenWeatherMap API.
The forecast includes temperature, humidity, and pressure data.
}
