% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plot.forecast}
\alias{plot.forecast}
\title{Plot Forecast Data}
\usage{
\method{plot}{forecast}(forecast_data, variable)
}
\arguments{
\item{forecast_data}{A dataframe or tibble of class `forecast` containing the forecasted weather data.}

\item{variable}{A character string indicating the variable to be plotted ("temperature", "humidity", or "pressure").}
}
\value{
A `ggplot` object displaying the weather forecast plot.
}
\description{
This function generates a plot for the 5-day weather forecast of a chosen variable (temperature, humidity, or pressure).
The plot includes a line, points, and an area shaded under the curve.
}
